% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRShiny.R
\name{MLRShiny}
\alias{MLRShiny}
\title{Launch MLRShiny Interface}
\usage{
MLRShiny()
}
\value{
Nothing
}
\description{
MLRShiny() loads interactive user interface built using R shiny.
}
\details{
Start MLRShiny

The interactive user interface is to provide an easy way for people who are learning Multiple Linear Regression. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
MLRShiny()
}
}
\keyword{MLRShiny}
